# api/src/main/java/io/wiiisdom/api/model/User.java
package io.wiiisdom.api.model;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class User {
    private String id;
    private String email;
    private String pseudo;
    private String password;
}