# api/src/main/java/io/wiiisdom/api/ressources/UserDto.java
package io.wiiisdom.api.ressources;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserDto {
    private String id;
    private String email;
    private String pseudo;
}